# specs/1-todo-evolution/contracts/task-api-contract.yaml
openapi: 3.0.0
info:
  title: Task Management API
  description: API for managing tasks in the Todo application
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
paths:
  /tasks:
    get:
      summary: Get all tasks for the authenticated user
      description: Retrieves a list of all tasks belonging to the authenticated user
      responses:
        '200':
          description: Successful response with list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}:
    get:
      summary: Get a specific task
      description: Retrieves details of a specific task by ID
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a task
      description: Updates details of an existing task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a task
      description: Deletes a specific task by ID
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/complete:
    patch:
      summary: Mark task as complete
      description: Marks a specific task as completed
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task marked as complete successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/incomplete:
    patch:
      summary: Mark task as incomplete
      description: Marks a specific task as incomplete
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task marked as incomplete successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - created_at
        - updated_at
        - user_id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        title:
          type: string
          maxLength: 255
          description: Title of the task
        description:
          type: string
          nullable: true
          description: Detailed description of the task
        completed:
          type: boolean
          description: Whether the task is completed
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated
        due_date:
          type: string
          format: date-time
          nullable: true
          description: When the task is due
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 1
          description: Priority level (1=low, 5=high)
        user_id:
          type: string
          format: uuid
          description: Owner of the task
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          description: Title of the task
        description:
          type: string
          nullable: true
          description: Detailed description of the task
        due_date:
          type: string
          format: date-time
          nullable: true
          description: When the task is due
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 1
          description: Priority level (1=low, 5=high)
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          description: New title of the task
        description:
          type: string
          nullable: true
          description: New description of the task
        due_date:
          type: string
          format: date-time
          nullable: true
          description: New due date for the task
        priority:
          type: integer
          minimum: 1
          maximum: 5
          description: New priority level (1=low, 5=high)
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message